add pytest with circleci (#98)
* add pytest with circleci

* add pytest in readme

* add requirements.txt and update readme

* Update README.md

* remove useless print from pytest lietensor and fix readme of imu preintegration
