+++ .circleci/config.yml
@@ -134,7 +134,27 @@ jobs:
                 \"pr_num\": \"${CIRCLE_PULL_REQUEST##*/}\"
               }
             }"
-
+  run_tests:
+    <<: *build_binary_params
+    docker:
+      - image: nvidia/cuda:<< parameters.cuda_version >>-cudnn<< parameters.cudnn_version >>-devel-ubuntu20.04
+    resource_class: small
+    steps:
+      - setup_env
+      - checkout
+      - attach_workspace:
+          at: .
+      - run:
+          name: Setup tests
+          command: |
+            export "PATH=~/miniconda3/bin:$PATH" && conda init bash && . $HOME/miniconda3/etc/profile.d/conda.sh
+            conda activate pyenv
+            pip install dist/pypose-*.whl
+            pip install -r requirements.txt
+            mkdir pytest-results
+            pytest --junitxml=pytest-results/junit.xml
+      - store_test_results:
+          path: pytest-results
 
 ##############################################################################
 # Workflows
@@ -148,6 +168,11 @@ workflows:
           torch_version: "1.11.0"
           python_version: "3.8"
       - build_docs:
+          requires: [linux_build_wheel]
+          cuda_version: "11.3.1"
+          torch_version: "1.11.0"
+          python_version: "3.8"
+      - run_tests:
           requires: [linux_build_wheel]
           cuda_version: "11.3.1"
           torch_version: "1.11.0"
